feat: add comments, address PR feedback, add more (skipped) tests
