+++ src/eventbridge/types.ts
@@ -1,0 +1,15 @@
+export interface EventBusRuleInput<
+  T = any,
+  DetailType extends string = string,
+  Source extends string = string
+> {
+  source: Source;
+  "detail-type": DetailType;
+  detail: T;
+  version: string;
+  id: string;
+  account: string;
+  time: string;
+  region: string;
+  resources: string[];
+}
